organising code into folders and moving all css into one file
