--- src/App.css
@@ -1,18 +1,0 @@
-.tab-list {
-  border-bottom: 2px solid #ccc;
-  padding-left: 0;
-  font-size: 1.5em;
-}
-
-.tab-list-item {
-  display: inline-block;
-  list-style: none;
-  margin-bottom: -2px;
-  padding: 0.5rem 0.75rem;
-}
-
-.tab-list-active {
-  background-color: white;
-  border: solid #ccc;
-  border-width: 2px 2px 0 2px;
-}
