+++ src/components/AddUser.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import axios from 'axios'
-import { Link, useHistory } from "react-router-dom";
+import { useHistory } from "react-router-dom";
 
 const AddUser = () => {
   let history = useHistory();
