+++ src/components/header.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
-import './header.css'
+
 class Header extends Component{
 render(){
   return (
-    <div className="Header">
+    <div className="header">
       <h1>Charing Cross Hospital Handover Portal</h1>
     </div>
   );
