--- src/header.css
@@ -1,12 +1,0 @@
-.Header{
-box-sizing: border-box;
-width: 100%;
-height: 125px;
-display: flex;
-justify-content: center;
-aligh-items: baseline;
-padding: 1em;
-margin-bottom: 2em;
-background-color: #33A2FF;
-color: #fff;
-}
