+++ src/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
+import 'bootstrap/dist/css/bootstrap.css';
+import './index.css';
 
 ReactDOM.render(
   <React.StrictMode>
